- 21-02-2005 fixed: iTunes sharing caused crash if no connection (or iTunes sharing) was available.
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk/XBMC@2889 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/FileSystem/FileDAAP.cpp
@@ -139,7 +139,7 @@ bool CFileDAAP::Open(const CURL& url, bool bBinary)
 		// If no host object returned then the connection failed
 		if (!m_thisHost) return false;		// tidy ups?
 
-		if (DAAP_ClientHost_Connect(m_thisHost) < 0) return false;		// tidy ups?
+		if ((int)DAAP_ClientHost_Connect(m_thisHost) < 0) return false;		// tidy ups?
 	}
 
 	/*
